requirements.txt
# Google ADK Framework
google-adk

# MCP (Model Context Protocol)
mcp

# LLM Integration
litellm

# Environment variables
python-dotenv